+++ Projet Orion/Orion/Orion/Carre.h
@@ -8,20 +8,20 @@ class Carre
 {
 public:
 	Carre();
-	Carre(double);
-	Carre(double, Point&);
-	Carre(double, Point&, Point&);
+	Carre(float);
+	Carre(float, Point&);
+	Carre(float, Point&, Point&);
 	virtual ~Carre();
 	Carre(const Carre &);
 	Carre& operator=(const Carre &);
 
 	void affiche() const;
 
-	double Getlongueur() const;
-	void Setlongueur(double);
+	float Getlongueur() const;
+	void Setlongueur(float);
 
 protected :
-	double m_longueur1, m_longueur2, m_longueur3, m_longueur4;
+	float m_longueur1, m_longueur2, m_longueur3, m_longueur4;
 	Point *m_M1;
 	Point *m_M2;
 };
